The cake was a lie
We probably could do this at the raw REST level, but it’s a giant pain
in the ass.  All the SDKs are in python or java, and R seems to work
better with java than with python.

So, we’re back to the aws sdk jar file.  It now has built-in support
for EMR and S3, so hopefully this one tiny, 2kb file will have all the
APIs we need.

It really would be nice if you could use:
1. An amazon security token
2. A JSON object
3. A REST API

To control and communicate with AWS.  Oh well, live and learn.